apply comments and move prune_lines to compose_network
