+++ scripts/gb_model/compose_network.py
@@ -31,6 +31,7 @@ from scripts.add_electricity import (
     attach_hydro,
     flatten,
 )
+from scripts.gb_model._helpers import get_lines
 
 logger = logging.getLogger(__name__)
 
@@ -832,6 +833,35 @@ def _prepare_costs(
     return costs
 
 
+def _prune_lines(
+    n: pypsa.Network,
+    prune_lines: list[dict[str, int]],
+) -> None:
+    """
+    Prune lines between bus0 and bus1 by setting their s_max_pu to 0.
+
+    Args:
+        n (pypsa.Network): The PyPSA network to modify.
+        prune_lines (list[dict[str, int]]): The lines to prune.
+    """
+    # Prune specified lines
+    for line in prune_lines:
+        mask = get_lines(n.lines, line["bus0"], line["bus1"])
+        if mask.any():
+            # Get lines to remove
+            lines_to_remove = n.lines[mask].index
+
+            # Remove the lines from the network
+            n.remove("Line", lines_to_remove)
+            logger.info(
+                f"Deleted line between bus {line['bus0']} and bus {line['bus1']}"
+            )
+        else:
+            logger.warning(
+                f"No line found to prune between bus {line['bus0']} and bus {line['bus1']}"
+            )
+
+
 def compose_network(
     network_path: str,
     output_path: str,
@@ -851,6 +881,7 @@ def compose_network(
     enable_chp: bool,
     ev_data: dict[str, str],
     ev_params: dict[str, float],
+    prune_lines: list[dict[str, int]],
 ) -> None:
     """
     Main composition function to create GB market model network.
@@ -897,6 +928,8 @@ def compose_network(
         Dictionary containing EV demand and flexibility data
     ev_params : dict[str, float]
         Dictionary containing EV profile adjustment parameters
+    prune_lines : list[dict[str, int]]
+        List of lines to prune between specified bus pairs
     """
     network = pypsa.Network(network_path)
     max_hours = electricity_config["max_hours"]
@@ -945,6 +978,8 @@ def compose_network(
     add_load(network, demands, eur_demand, year)
 
     add_EVs(network, ev_data, ev_params, year)
+
+    _prune_lines(network, prune_lines)
 
     finalise_composed_network(network, context)
 
@@ -1011,4 +1046,5 @@ if __name__ == "__main__":
         enable_chp=snakemake.params.enable_chp,
         ev_data=ev_data,
         ev_params=ev_params,
+        prune_lines=snakemake.params.prune_lines,
     )
